UPDATE node 18.2.0 (#3825)
* UPDATE node 18.2.0

* FIX node.js v18.2.0 problems

* FIX test for error

* UPDATE dexie js

* UPDATE react

* UPDATE react-scripts

* UPDATE node to 18.3.0

* TRY fix react-native example

* FIX ci

* FIX ci again

+++ test/unit/replication-graphql.test.ts
@@ -39,6 +39,7 @@ import {
 import {
     getLastPullDocument,
     getLastPushCheckpoint,
+    RxReplicationError,
     setLastPullDocument
 } from '../../plugins/replication';
 import * as schemas from '../helper/schemas';
@@ -1413,7 +1414,10 @@ describe('replication-graphql.test.ts', () => {
                     first()
                 ).toPromise();
 
-                assert.ok(ensureNotFalsy(error).toString().includes('foobar'));
+                if (!error || (error as RxReplicationError<any>).type !== 'pull') {
+                    console.dir(error);
+                    throw error;
+                }
 
                 replicationState.cancel();
                 c.database.destroy();
